Feat: 존재하지 않는 PK 삭제 시, 예외 발생하도록 변경
- 삭제 성공/실패 여부를 프론트엔드측이 파악하도록 하기 위함

+++ checking/src/main/java/sheetplus/checkings/domain/service/EntryCRUDService.java
@@ -75,7 +75,11 @@ public class EntryCRUDService {
 
     @Transactional
     public void deleteEntry(Long id){
-        entryRepository.deleteById(id);
+        if(entryRepository.existsById(id)){
+            entryRepository.deleteById(id);
+        }else{
+            throw new ApiException(ENTRY_NOT_FOUND);
+        }
     }
